New iupStrToDouble and iupStrReturnDouble
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@2223 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/srcscintilla/iupsci_style.c
@@ -177,7 +177,7 @@ static char* iScintillaGetFontSizeFracStyleAttrib(Ihandle* ih, int style)
 
   size = iupScintillaSendMessage(ih, SCI_STYLESETSIZEFRACTIONAL, style, 0);
 
-  return iupStrReturnFloat((float)size / SC_FONT_SIZE_MULTIPLIER);
+  return iupStrReturnFloat((float)size/(float)SC_FONT_SIZE_MULTIPLIER);
 }
 
 static int iScintillaSetFontSizeFracStyleAttrib(Ihandle* ih, int style, const char* value)
